[WCBRD-1589] Fix coverity issue

+++ src/platforms/rw61x/mcux_els.c
@@ -105,14 +105,14 @@ static status_t ELS_PRNG_KickOff(void)
         for(keyIdx = 0; keyIdx < MCUXCLELS_KEY_SLOTS; keyIdx++)
         {
             /* find a free key slot in ELS keystore */
-            status = ELS_check_key(keyIdx, &key_properties);
+            status = ELS_check_key((uint8_t)keyIdx, &key_properties);
             if (status != kStatus_Success)
             {
                 return kStatus_SlotUnavailable;
             }   
             
             /* Found free key slot */
-            if(key_properties.bits.kactv == 0)
+            if(key_properties.bits.kactv == 0u)
             {
                 break; 
             }
